+++ COMMON/headers/TXTreader.h
@@ -2,7 +2,14 @@
 #define TXT_READER_H
 
 #include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <ctype.h>
+#include "IsBadPtr.h"
+#include "SizeFile.h"
+#include "LineCounter.h"
 
-char** TXTreader(FILE *SourceFile, char* buffer, size_t *len_buffer, size_t *count_line, int (*convert)(int));
+
+char** TXTreader(FILE *SourceFile, char* buffer, size_t *len_buffer, int *count_line, int (*convert)(int));
 
 #endif
